feat(eco): Wire real-byte tracking from downloads + web search
HttpModelDownloadManager records a networkEvent per download session
(delta bytes, not total file size, so resumes don't double-count).
Partial-failure rows are preserved for honest accounting.

WebSearchExecutor records request + response bytes per DuckDuckGo
round-trip. Session ID "websearch" groups the events.

Both accept EcoMetricsRepository as an optional constructor param —
null = skip tracking, used by tests and isolated builds. Koin wires
the live repository on Android.

Widened EcoSavings.bytesSent / bytesReceived from Int to Long. Big
M1K3 downloads are ~2.4 GB which overflows Int.MAX_VALUE. Test call
sites updated with L suffix.

MurphySig: kev+claude / confidence 0.82 / 2026-04-19

+++ app/composeApp/src/commonMain/kotlin/app/m1k3/ai/assistant/eco/EcoCalculator.kt
@@ -73,8 +73,8 @@ object EcoCalculator {
             waterSavedMl = (WATER_ML_PER_100_TOKENS * scaleFactor).toInt(),
             energySavedWh = (ENERGY_MWH_PER_100_TOKENS * scaleFactor).toInt(),
             co2PreventedG = (CO2_G_PER_100_TOKENS * scaleFactor).toInt(),
-            bytesSent = 0,
-            bytesReceived = 0,
+            bytesSent = 0L,
+            bytesReceived = 0L,
         )
     }
 
@@ -86,10 +86,10 @@ object EcoCalculator {
      * "cloud bytes avoided" headline can be derived from a single table.
      */
     fun networkEvent(
-        bytesSent: Int,
-        bytesReceived: Int,
+        bytesSent: Long,
+        bytesReceived: Long,
     ): EcoSavings {
-        require(bytesSent >= 0 && bytesReceived >= 0) { "Bytes must be non-negative" }
+        require(bytesSent >= 0L && bytesReceived >= 0L) { "Bytes must be non-negative" }
         return EcoSavings(
             tokensProcessed = 0,
             waterSavedMl = 0,
@@ -205,8 +205,8 @@ data class EcoSavings(
     val waterSavedMl: Int,
     val energySavedWh: Int, // Stored as milliwatt-hours (mWh)
     val co2PreventedG: Int,
-    val bytesSent: Int = 0,
-    val bytesReceived: Int = 0,
+    val bytesSent: Long = 0L,
+    val bytesReceived: Long = 0L,
 )
 
 /**
